[CLN] Controller 합치기, RestController 이름 Api prefix로 변경
Resolves #165
