--- src/main/java/kr/ac/sejong/web/CourseScheduleController.java
@@ -1,17 +1,0 @@
-package kr.ac.sejong.web;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-
-@Controller
-@RequiredArgsConstructor
-public class CourseScheduleController {
-
-    @GetMapping(path = "/trackSubject")
-    public String trackSubject() {
-        return "trackSubject";
-    }
-}
-
