--- src/main/java/kr/ac/sejong/web/TrackAllController.java
@@ -1,13 +1,0 @@
-package kr.ac.sejong.web;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class TrackAllController {
-
-    @GetMapping("/trackAll")
-    public void trackAll() throws Exception{
-
-    }
-}
