+++ src/main/java/kr/ac/sejong/web/exception/CommonExceptionAdvice.java
@@ -1,4 +1,4 @@
-package kr.ac.sejong.web;
+package kr.ac.sejong.web.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
